requirements.txt
Flask
opencv-python
numpy
ultralytics
google-auth
flask-mail
Werkzeug